backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-Mail==0.9.1
Werkzeug==2.3.7
python-socketio==5.8.0
python-engineio==4.7.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.33.3
pytz==2023.3
